requirements.txt
fastapi
uvicorn
python-dotenv
llama-index-core>=0.12.0
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
llama-index-llms-bedrock
llama-index-postprocessor-colbert-rerank
qdrant-client
transformers
torch
pydantic
llama-index-core 
llama-index-embeddings-huggingface 
llama-index-readers-file 
python-dotenv
llama-index-vector-stores-qdrant 
llama-index-readers-file 
llama-index-embeddings-fastembed 
llama-index-llms-openai
termcolor
llama-index-llms-bedrock
llama-index-postprocessor-colbert-rerank
transformers 
torch
fastapi
uvicorn